+++ webconsole.client/client/client/app/components/services/services.component.js
@@ -1,4 +1,4 @@
-System.register(['@angular/core', '@angular/common', '../../services/backend.service', '../../services/appglobals.service', "@angular/router"], function(exports_1, context_1) {
+System.register(['@angular/core', '@angular/common', '../../services/backend.service', '../../services/appglobals.service', "@angular/router", '../../pipes/servicesFilter.pipe'], function(exports_1, context_1) {
     "use strict";
     var __moduleName = context_1 && context_1.id;
     var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
@@ -10,7 +10,7 @@ System.register(['@angular/core', '@angular/common', '../../services/backend.ser
     var __metadata = (this && this.__metadata) || function (k, v) {
         if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
     };
-    var core_1, common_1, backend_service_1, appglobals_service_1, router_1;
+    var core_1, common_1, backend_service_1, appglobals_service_1, router_1, servicesFilter_pipe_1;
     var ServicesComponent;
     return {
         setters:[
@@ -28,6 +28,9 @@ System.register(['@angular/core', '@angular/common', '../../services/backend.ser
             },
             function (router_1_1) {
                 router_1 = router_1_1;
+            },
+            function (servicesFilter_pipe_1_1) {
+                servicesFilter_pipe_1 = servicesFilter_pipe_1_1;
             }],
         execute: function() {
             ServicesComponent = (function () {
@@ -35,6 +38,7 @@ System.register(['@angular/core', '@angular/common', '../../services/backend.ser
                     this.router = router;
                     this._backend = _backend;
                     this._appGlobals = _appGlobals;
+                    this.searchName = '';
                 }
                 ServicesComponent.prototype.onSelect = function (serviceId) {
                     this.router.navigate(['/services', serviceId]);
@@ -57,6 +61,7 @@ System.register(['@angular/core', '@angular/common', '../../services/backend.ser
                         directives: [common_1.FORM_DIRECTIVES, common_1.NgFor, common_1.NgFormModel],
                         providers: [backend_service_1.BackendServices],
                         templateUrl: 'app/html/services/services.template.html',
+                        pipes: [servicesFilter_pipe_1.ServicesFilter],
                     }), 
                     __metadata('design:paramtypes', [router_1.Router, backend_service_1.BackendServices, appglobals_service_1.AppGlobals])
                 ], ServicesComponent);