+++ webconsole.client/client/src/app/components/services/services.component.ts
@@ -7,6 +7,8 @@ import {Service} from '../../domain/service';
 import {Bundle} from '../../domain/bundle';
 import {ROUTER_DIRECTIVES, Router} from "@angular/router";
 
+import {ServicesFilter} from '../../pipes/servicesFilter.pipe'
+
 //import {BundlesService} from '../services/bundles.service';
 
 declare var jQuery: any;
@@ -16,11 +18,14 @@ declare var jQuery: any;
     directives: [FORM_DIRECTIVES, NgFor, NgFormModel],
     providers: [BackendServices],
     templateUrl: 'app/html/services/services.template.html',
+    pipes: [ServicesFilter],
     //styleUrls:  ['app/js/datatables.css']
 })
 export class ServicesComponent implements OnInit {
 
     services: Service[];
+    public searchName: string = '';
+
 
     constructor(private router: Router, private _backend: BackendServices, private _appGlobals: AppGlobals) {
     }