+++ webconsole.it/src/io/skysail/webconsole/it/BackendBundlesTest.java
@@ -4,6 +4,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.FrameworkUtil;
@@ -26,6 +27,7 @@ public class BackendBundlesTest {
     }
 
     @Test
+    @Ignore
     public void testName() throws Exception {
         String bundlespage = browser.getBundles();
         System.out.println(bundlespage);